### 2019-01-03 #1152 @pankajkmishra
Hi @kyleniemeyer - I think @Kevin-Mattheus-Moerman ( already in the JOSS list) could be a potential reviewer.
