### 2019-01-03 #1152 @Kevin-Mattheus-Moerman
I can review this, looks interesting.
