### 2019-02-26 #1173 @vijaysm
@kyleniemeyer Should I upload a list of my suggestions separately or just continue discussion with the author here as part of the comment thread ?

@pankajkmishra  I also noticed that there is no official release for NodeLab yet. So it may be appropriate to update the versioning info first and then tag the commit as a checkpoint for the review. Thoughts @kyleniemeyer ?
