### 2019-02-26 #1173 @kyleniemeyer
@vijaysm feel free to continue the discussion here; you can submit a list of suggestions as a single comment and we can go from there.

Regarding the official release, I'm fine if @pankajkmishra waits until after making any changes as part of this review process, and then makes an official release associated with the version accepted here.
