### 2019-03-07 #1173 @Kevin-Mattheus-Moerman
Comments on paper:

- [ ] Add "be" in the following: "....(Persson & Strang, 2004), which can [be] computed based on a..."

- [ ] Remove plural in: "As a result~s~, NodeLab can take the following..."

- [ ] Rephrase "some discrete set of point cloud on the boundary, which need not to be uniformly
sampled." e.g. to: "some discrete ~set of~ point cloud on the boundary, which need not ~to~ be uniformly
sampled."

- [ ] Can you work to rephrase/clarify this awkward sounding sentence: "The boundary can be smoothed through curve-interpolation according to the fill-distance near the boundary, which provides the flexibility to create the domain by manually digitizing of the geometry from a hand-drawing, digital-drawing, or a downloaded image."

- [ ] Can you add  a References heading?
