### 2019-03-07 cbf8b13
Updated demo to double as publishable documentation
