### 2019-03-07 #1173 @Kevin-Mattheus-Moerman
I've created a pull request (https://github.com/pankajkmishra/NodeLab/pull/2) to enhance the demos/documentation. I recommend clarifying the required input arguments and perhaps to visualize them as I suggest.
I also recommend making use of the MATLAB documentation generation tools. I give an example in that demo. You can publish it to create HTML documentation which can be integrated in MATLAB.
