+++ paper.bib
@@ -7,7 +7,7 @@
   number={7},
   pages={531--544},
   year={2015},
-  doi={10.1016/j.camwa.2015.01.009}
+  doi={10.1016/j.camwa.2015.01.009},
   publisher={Elsevier}
 }
 
@@ -19,7 +19,7 @@
   number={2},
   pages={329--345},
   year={2004},
-  doi={10.1137/s0036144503429121}
+  doi={10.1137/s0036144503429121},
   publisher={SIAM}
 }
 
