### 2019-03-08 #1173 @arfon
@pankajkmishra - there were some missing commas in your bibtex: https://github.com/pankajkmishra/NodeLab/pull/3
