### 2019-03-11 6d0d49f
Improved demos
I have added more descriptions in the demos. I'll add a separate section in the 'readme' file explaning different input and output variables of NodeLab
