### 2019-03-11 cb232fb
updated the paper
updated the paper using the reviewer's suggestion
