### 2019-03-11 #1173 @pankajkmishra
> Hi @Kevin-Mattheus-Moerman ! I have updated the paper according to the following comments.  I have also merged your expanded visualization for the first test, which would (indeed) provide more insight to the user. However, I have not included this visualization to other tests to keep the demo's simple and easy to integrate in external (meshfree) PDE solvers.  I have added the description of Input/Output variables in the 'readme' file.
>
>     * [x]  Add "be" in the following: "....(Persson & Strang, 2004), which can [be] computed based on a..."
>
>     * [x]  Remove plural in: "As a results, NodeLab can take the following..."
>
>     * [x]  Rephrase "some discrete set of point cloud on the boundary, which need not to be uniformly
>       sampled." e.g. to: "some discrete set of point cloud on the boundary, which need not to be uniformly
>       sampled."
>
>     * [x]  Can you work to rephrase/clarify this awkward sounding sentence: "The boundary can be smoothed through curve-interpolation according to the fill-distance near the boundary, which provides the flexibility to create the domain by manually digitizing of the geometry from a hand-drawing, digital-drawing, or a downloaded image."
>
>     * [x]  Can you add  a References heading?
