### 2019-03-11 #1173 @pankajkmishra
@kyleniemeyer @vijaysm : The release version is v1.0 after incorporating @Kevin-Mattheus-Moerman 's suggestions.
