### 2019-03-11 #1173 @Kevin-Mattheus-Moerman
@pankajkmishra

Thanks for implementing most of the points. Below are some remaining points on the paper:

- [ ] This part has not been changed yet: "discrete set of point cloud". Should this be "discrete sets of point clouds" or "discrete point cloud sets" ? Also "which need not to be" sounds odd, I suggest you remove the word "to".

- [ ] Should the long hyphen be removed before state of the art? "... (and keep) it -state-of-the-art by incorporating..."
