### 2019-03-11 #1173 @Kevin-Mattheus-Moerman
@pankajkmishra thanks for expanding the demo file.

Minor point, perhaps update this comment:
```matlab
% You can add similar visualization section for other demos, or your own experiment
% to have a better understanding of NodeLab.
```
To:
```matlab
% You can add similar visualization sections to other demos or your own code
% to have a better understanding of the inputs used in NodeLab.
```
