### 2019-03-11 #1173 @Kevin-Mattheus-Moerman
@pankajkmishra about the license. The GNU license you've chosen, strictly speaking, requires that you add the following "boiler plate text" under all your codes (see also: https://opensource.org/licenses/GPL-3.0):

```
    <one line to give the program's name and a brief idea of what it does.>
    Copyright (C) <year>  <name of author>

    This program is free software: you can redistribute it and/or modify
    it under the terms of the GNU General Public License as published by
    the Free Software Foundation, either version 3 of the License, or
    (at your option) any later version.

    This program is distributed in the hope that it will be useful,
    but WITHOUT ANY WARRANTY; without even the implied warranty of
    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
    GNU General Public License for more details.

    You should have received a copy of the GNU General Public License
    along with this program.  If not, see <http://www.gnu.org/licenses/>.
```
 I recommend you update this text with your information and paste it at the end of your codes. If you want to automate this, check out how I did this for my personal project [GIBBON](https://github.com/gibbonCode/GIBBON).  I use an automated script to remove or add these boiler plates to the bottom of selected functions (see [addCodeFooter.m](https://github.com/gibbonCode/GIBBON/blob/master/docs/docTools/addCodeFooter.m) and [removeCodeFooter.m](https://github.com/gibbonCode/GIBBON/blob/master/docs/docTools/removeCodeFooter.m))
