### 2019-03-11 #1173 @pankajkmishra
@Kevin-Mattheus-Moerman Done
> ```matlab
> % You can add similar visualization section for other demos, or your own experiment
> % to have a better understanding of NodeLab.
> ```
>
> To:
>
> ```matlab
> % You can add similar visualization sections to other demos or your own code
> % to have a better understanding of the inputs used in NodeLab.
> ```
