### 2019-03-11 #1173 @Kevin-Mattheus-Moerman
Nearly there with the paper. I meant to remove the hyphen altogether, so use "state-of-the-art" instead of "-state-of-the-art".
