### 2019-03-11 #1173 @pankajkmishra
> Nearly there with the paper. I meant to remove the hyphen altogether, so use "state-of-the-art" instead of "-state-of-the-art".

Done.
