### 2019-03-11 7e18ff7
created run_all_tests
Sets up and runs all the tests in the 'demos' directory.
