+++ run_all_tests.m
@@ -1,0 +1,24 @@
+setup 
+cd demos
+test_1C
+disp('Press any key to run the next test!')
+pause;
+clear variables; close all
+test_2R
+disp('Press any key to run the next test!')
+pause;
+clear variables; close all
+test_2Lshape
+disp('Press any key to run the next test!')
+pause;
+clear variables; close all
+test_3Lake
+disp('Press any key to run the next test!')
+pause;
+clear variables; close all
+test_4LakeIsland
+disp('All tests are complete. Press any key exit')
+pause;
+clear variables; close all
+
+cd ..
