### 2019-03-11 #1173 @pankajkmishra
> @pankajkmishra I've left the "automated test" box above unticked. I think manually running the codes in the demos folder is fine, so the following is a recommendation rather than a requirement.
>
> To add a basic type of automated test functionality, I recommend you just create a single script called `testNodeLab` in the main folder which runs all the demos. (I have something similar, but a bit more complicated, for GIBBON here: https://github.com/gibbonCode/GIBBON/blob/master/lib/testGibbon.m).

I have added a simple "automated test", which sets up the path and runs all the tests in the demos directory one-by-one. Check it out.
