### 2019-03-11 316ea3b
Update run_all_tests.m
