### 2019-03-11 #1173 @Kevin-Mattheus-Moerman
@pankajkmishra can you describe in more detail what advances/novelty your work offers over the work (and MATLAB script) by [Fornberg & Flyer](https://www.sciencedirect.com/science/article/pii/S0898122115000334). If one glances at your project it seems, as stated in the paper, that the core of the matter relies on `node_places` which appears to stem rather directly from the Fornberg & Flyer implementation, and that on top of that your library contains about 6 small scripts containing between 5-15 lines of code which appear to simply call the external codes from the ext folder. Therefore, in order to understand/determine if this work is a significant advance on those published works, can you describe what changes you've made and what technical novelty you've added/enabled?
