### 2019-03-11 #1173 @pankajkmishra
@Kevin-Mattheus-Moerman : The main idea is to integrate the node_placing algorithm in a geometric modeling framework like FEM mesh-generators.

I am using Fornberg & Flyers approach for initial node-generation in the bounding box (rectangle) and then using the geometric modeling method similar to what (Persson & Strang, 2004) do for generating FEM meshes - but not using the iterative (time-consuming part) refinement, because meshless methods does not need as perfect nodes as FEM. I am using an external C++ code for  creating polygon with point-clouds and a curve interpolation function to keep the user defined density of boundary nodes. I initially had written this for my own numerical tests and thought that it might be useful to others working in meshfree community.

As far as the value addition over Forberg & Flyer is concerned - If you see page 24 - Table 2 (first column) in (https://arxiv.org/pdf/1812.03160.pdf) they mention a few limitations of Fornberg & Flyer (FF) approach

(1) FF supports partial irregular domains while here we can create nodes in any kind of irregular domains.
(2) FF lacks compatibility with boundary nodes - the present work does not (providing manual control)
(3) FF does not guarantee 'minimal spacing' between nodes - however in the present work we ensure the minimal spacing through the variable 'ptol'.

I had the feeling from JOSS author guidelines that I shouldn't mention these details.
