### 2019-03-12 #1173 @pankajkmishra
@Kevin-Mattheus-Moerman @kyleniemeyer : If this work does not fit in JOSS's aim and scope, feel free to let me know. I'd be fine to keep it 'just as a github repository', and others can still use it, which was the original purpose.
