### 2019-03-31 #1173 @kyleniemeyer
@Kevin-Mattheus-Moerman does the explanation above satisfy your concerns?

@pankajkmishra It may be helpful to specifically differentiate your work from the F&F paper and mention how you build on that. Also, any code in `ext` that came from somewhere else should be cited in the paper.
