### 2019-05-07 #1173 @vijaysm
@pankajkmishra Please document that the statistics toolbox is a required dependency in order to compute local pairwise distances between points. Instead, this dependency can be easily removed by exposing your own `pdist2` implementation like below.
```
function [d]=pdist2loc(cpts,xyp)
  d=sqrt((cpts(:,1)-xyp(:,1)).^2+(cpts(:,2)-xyp(:,2)).^2);
end
```
and then using `pdist2loc` instead of `pdist2` in the tests:
```
radius = @(p,ctps) 0.005+0.08*(min(pdist2loc(ctps, p))); % for variable node-density
```
