### 2019-05-11 #1173 @vijaysm
@pankajkmishra It would also be useful to talk about future extensions to 3-D domains and a simple tessellation algorithm with the node placement such that an unstructured mesh can be generated with user-specified refinement criteria. This would improve the application of the NodeLab software to not just meshfree solvers, but even for adaptive FEM simulations. This is more of a suggestion for adding a TODO file to the repository with some extension ideas.
