### 2019-05-15 #1173 @Kevin-Mattheus-Moerman
@kyleniemeyer in relation to
> @Kevin-Mattheus-Moerman does the explanation above satisfy your concerns?

Yes, somewhat. My key "concern" is that this submission is in my opinion borderline in terms of what it adds on top of previous work, and very close to being too minor a contribution. This work appears to largely be a collection of codes (with minor changes) developed by others, combined with a small amount of novel codes. Despite the above, I have ticked all the boxes, since I do feel this is a very useful project, and the author has added useful examples and demos. If the editor @kyleniemeyer and other reviewer @vijaysm feel favourable about this submission I am happy to accept this work.
