### 2019-05-16 #1173 @pankajkmishra
Thank you @vijaysm - I'll edit the repository according to your suggestions. I have already mentioned 3D in the future development. However, I am not targeting FEM users at the moment.

I do agree with @Kevin-Mattheus-Moerman remark about using mostly external codes. I could have rewritten those codes (e.g. FF's node placing, or desegment.cpp) and added it to the library, which would, probably,  require extra efforts from a user (who is already familiar with the external codes I have used) to understand my repository.

I understand that it looks "borderline" in terms of code writing, but I do believe (as the first user) that it is useful to the target users because the end results, which is not that "borderline". Each external and internal code other than "node_placing.m" adds something to it. Moreover, The fact that there is a lack of open-source repository for meshfree node-generation makes this repository useful at the moment.

@kyleniemeyer I have cited all the external sources. I'll wait for your decision on this and then do the final polishing to the repository.

Best regards,
