### 2019-06-08 #1173 @labarba
@vijaysm — What is your assessment with regards to the [issues raised](https://github.com/openjournals/joss-reviews/issues/1173#issuecomment-492571822) by @Kevin-Mattheus-Moerman? What about you, @kyleniemeyer? The author has received no reply for 23 days. Do you feel this can go on to publication, or is the "minor contribution" assessment a blocker?
