### 2019-06-09 #1173 @vijaysm
@labarba @kyleniemeyer I agree with @Kevin-Mattheus-Moerman that the contribution is an incremental update to the work by Fornberg [1] in terms of the algorithmic aspects. However, I feel that NodeLab exposes the node-generation infrastructure in a general setting that is directly usable by solvers written in MATLAB. Extensions to 3-D would have made a stronger case for acceptance of the current submission, especially given the fact that a new preprint by Sande and Fornberg is now available on Arxiv [2].

Given the potential for this contribution to support practical use-cases, I can provide a conditional acceptance if the following comments are addressed:

- A more detailed discussion on how NodeLab builds on the original F&F contribution,
- Handling node generation for 3-D domains with references to [2],
- Notes on future extensions to arbitrary irregular domains

[1] Fornberg, B.,  Flyer, N. (2015). Fast generation of 2-D node distributions for mesh-free PDE discretizations. Computers & Mathematics with Applications, 69(7), 531-544.
[2] Kiera van der Sande, Bengt Fornberg, Fast variable density 3-D node generation, https://arxiv.org/abs/1906.00636.
