### 2019-06-10 #1173 @pankajkmishra
Thank you! I'll respond to @vijaysm 's comments ASAP. I switched job and moved from Hong Kong to US.
