### 2019-06-24 #1173 @kyleniemeyer
@pankajkmishra no worries—I'm going to set an automated reminder for another week from now.
