+++ paper.md
@@ -19,7 +19,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-Meshless or meshfree methods, which work globally or locally are up-and-coming numerical methods for numerical modeling, which unlike other mesh-based algorithms, do not require an structured grid/elements in the domain. These methods can work with scattered nodes in the domain, however, the nature of the modeling problem often suggests a certain type of node-layout to get the most efficient modeling results.
+Meshless/meshfree methods are cutting-edge numerical methods for numerical modeling, which unlike other mesh-based algorithms, do not require an structured grid/elements in the domain. These methods can work with scattered nodes in the domain, however, the nature of the modeling problem often suggests a certain type of node-layout to get the most efficient modeling results. Many real-world applications (like weather forecasting, tsunami modeling, and geophysical imaging, computational mechanics) require to generate a large number of nodes in a significantly large arbitrary domain. The literature in this context is up-and-coming and some recent progress include: (Fornberg and Flyer, 2015; Michaels, 2017; Shankar et al., 2018; Vlasiuk et al., 2017; Zamolo and Nobile, 2018), which focus on different aspects of node-generation based on typical requirements. The node-placing approach by Fornberg and Flyer (2015) is similar to advancing front methods and has been reported to have advantages like computational speed, simple algorithms, and good quality of distribution. 
 
 ``NodeLab`` is a simple MATLAB-repository for node-generation and adaptive refinement for testing, and implementing various meshfree methods for solving PDEs in arbitrary domains. The core-algorithm behind this package is the *node-placing* [@Fornberg:2015] approach because of its simplicity, computational speed and the quality of the distribution. The *node-placing* method has been used for creating initial node-distribution in the \emph{bounding-box} of the desired domain. To represent the geometry of the domain, NodeLab couples the Signed-distance function (SDF) [@Persson:2004], which can be computed based on *a priory* information about the domain. As a result, NodeLab can take the following geometric objects as inputs: 
 
