### 2019-08-02 ef60545
Update paper.md
