+++ paper.md
@@ -19,16 +19,9 @@ bibliography: paper.bib
 ---
 
 # Summary
-Meshless/meshfree methods are cutting-edge numerical methods for numerical modeling, which unlike other mesh-based algorithms, do not require an structured grid/elements in the domain. These methods can work with scattered nodes in the domain, however, the nature of the modeling problem often suggests a certain type of node-layout to get the most efficient modeling results. Many real-world applications (like weather forecasting, tsunami modeling, and geophysical imaging, computational mechanics) require to generate a large number of nodes in a significantly large arbitrary domain. The literature in this context is up-and-coming and some recent progress include: (Fornberg and Flyer, 2015; Michaels, 2017; Shankar et al., 2018; Vlasiuk et al., 2017; Zamolo and Nobile, 2018), which focus on different aspects of node-generation based on typical requirements. The node-placing approach by Fornberg and Flyer (2015) is similar to advancing front methods and has been reported to have advantages like computational speed, simple algorithms, and good quality of distribution. 
+Meshless/meshfree methods are cutting-edge numerical methods for numerical modeling, which unlike other mesh-based algorithms, do not require an structured grid/elements in the domain. These methods can work with scattered nodes in the domain, however, the nature of the modeling problem often suggests a certain type of node-layout to get the most efficient modeling results. Many real-world applications (like weather forecasting, tsunami modeling, and geophysical imaging, computational mechanics) require to generate a large number of nodes in a significantly large arbitrary domain. The literature in this context is up-and-coming, which focus on different aspects of node-generation based on typical requirements. The node-placing approach by [@Fornberg:2015] is similar to advancing front methods and has been reported to have advantages like computational speed, simple algorithms, and good quality of distribution. 
 
-``NodeLab`` is a simple MATLAB-repository for node-generation and adaptive refinement for testing, and implementing various meshfree methods for solving PDEs in arbitrary domains. The core-algorithm behind this package is the *node-placing* [@Fornberg:2015] approach because of its simplicity, computational speed and the quality of the distribution. The *node-placing* method has been used for creating initial node-distribution in the \emph{bounding-box} of the desired domain. To represent the geometry of the domain, NodeLab couples the Signed-distance function (SDF) [@Persson:2004], which can be computed based on *a priory* information about the domain. As a result, NodeLab can take the following geometric objects as inputs: 
-
-* simple shapes (like rectangles, and circles), required to create the desired geometry. For example, the SDF for a typical *plate with a hole* can be computed through a rectangle and a circle. 
-* a level-set representation $D(x,y)=0$ of the domain, or 
-* some discrete sets of point clouds on the boundary, which need not be uniformly
-sampled. NodeLab uses curve-interpolation to redisctribute the input point-cloud as per user-defined density, which provides the flexibility to create the domain by manually digitizing of the geometry from any image.
-
-The node-distribution in ``NodeLab`` can be  refined non-uniformly by adapting the information provided through *control-points*, which are an input from the user. *control-points* provide spatial locations where the user needs relatively finer nodes. 
[…]
