### 2019-08-02 dd85881
Update paper.md
