+++ readme.md
@@ -1,4 +1,4 @@
-## NodeLab --- REVISING .. Will Finish by End of July, 2019
+## NodeLab
 NodeLab is a simple MATLAB-package for unstructured node-generation and refinement for meshfree modeling in arbitrary domains. The final can be seen in the 'workspace'. Following are the description of Input/Output variables of NodeLab:
 ### Input: 
   * *box* - a vector defining the bounding box of the domain.
