+++ readme.md
@@ -1,5 +1,5 @@
 ## NodeLab
-NodeLab is a simple MATLAB-package for unstructured node-generation and refinement for meshfree modeling in arbitrary domains. The final can be seen in the 'workspace'. Following are the description of Input/Output variables of NodeLab:
+NodeLab is a simple MATLAB-package for unstructured node-generation and refinement for meshfree modeling in 2D arbitrary domains. The final output can be seen in the 'workspace'. Following are the description of Input/Output variables of NodeLab:
 ### Input: 
   * *box* - a vector defining the bounding box of the domain.
   * *hbdy* - the density of the boundary poit-cloud
