+++ readme.md
@@ -15,6 +15,7 @@ NodeLab is a simple MATLAB-package for unstructured node-generation and refineme
 * Open MATLAB
 * Go to the directory 'NodeLab'
 * run the script 'setup.m'. This will add different directories of the NodeLab on your MATLAB path — only for the current session. Now you are ready to run the demos and generate nodes for your own project. 
+Note: make sure you have MATLAB's statistics toolbox. 
 
 ## Tutorials
 ## 1.1
