### 2019-08-02 #1173 @pankajkmishra
@kyleniemeyer I have updated the paper showing better clarity for the value addition:
Added the statistics toolbox  dependency as @vijaysm had suggested.  However, I couldn't make a similar 3D extension of it at the moment and it continues to be in the future work section. If the present form of the repository does not qualify to be listed in JOSS, I would be happy to withdraw the submission :)
P.S. I have a paper under review where I need to cite this work, which was the original purpose of submitting this to the JOSS.
