### 2019-08-04 #1173 @vijaysm
@pankajkmishra Thanks for the description update.

@kyleniemeyer While the updated paper makes the application space clearer than before, I would like to reiterate that the work on NodeLab still remains incremental. Node generation support for 3-D volumes or surfaces would certainly compel me to accept the paper. Also, other open-source Matlab packages by Per-Olof Persson (Distmesh [1]) and by Darren Engwirda (Mesh2D [2]), already do provide adaptive node generation and delaunay triangulation based on user-defined distance functions as well for 2-D domains. Additionally, I think some NodeLab examples showing adaptive refinement with integration to a mesh-free solver would have added value to this contribution with a practical use-case.

As it stands, I am taking the decision to **reject** the paper submission. @pankajkmishra, please consider updating examples and adding support for 3-D domains to differentiate this work from other existing open-source alternatives. I would be happy to discuss my comments further if you have any questions.

[1] http://persson.berkeley.edu/distmesh/
[2] https://github.com/dengwirda/mesh2d
