### 2019-08-04 #1173 @kyleniemeyer
Hi @vijaysm, thanks for your input.

I should point out that JOSS does not reject submissions for novelty, originality, or impact; instead, we only reject when the authors are unwilling to improve the software to meet our standards. Since both you and @Kevin-Mattheus-Moerman have checked off all the review items, and it doesn't seem like there are any remaining items of that nature, I am going to move this into the final acceptance stages.

@pankajkmishra at this point please archive the software repository (e.g., in Zenodo) and provide the DOI here. I'm going to do a final review/edit of the paper itself.
