### 2019-08-04 #1173 @vijaysm
@kyleniemeyer Fair enough. In that case, the description update from @pankajkmishra was all that was needed to make the submission complete for me.
