### 2019-08-12 #1173 @kyleniemeyer
Hi @pankajkmishra , just a few final edits needed on the paper:

- can you provide more-complete affiliation info, like what you'd typically see in a journal article? (i.e., city, country)
- Please merge my PR for fixes to the bib file: https://github.com/pankajkmishra/NodeLab/pull/4
- I did some copy editing to the paper in https://github.com/pankajkmishra/NodeLab/pull/5, please merge that.

There are a few more places in the paper that should be edited:
> The literature in this context is up-and-coming, which focus on different aspects of node-generation based on typical requirements. The node-placing approach by @Fornberg:2015 is similar to advancing front methods and has been reported to have advantages like computational speed, simple algorithms, and good quality of distribution.
- The first sentence here doesn't make sense to me—I'm not sure what it's actually saying.
- Anytime you reference "literature", it should be followed by multiple citations. Also, what are the "typical requirements"?
- You mention "advancing front methods", but provide no definition or citation. Why is the fact that the node-placing approach is similar to these methods meaningful?

Aside from that, the first paragraph should start with a statement of need, meant for a general reader. I think you could move your sentence about the applications here, and do some revision to make it work.
