### 2019-08-14 12d3c82
Update paper.md
