### 2019-08-14 b0f6238
Update paper.md
