### 2019-08-14 f1c656d
Update paper.md
