### 2019-08-15 1565c43
Delete NodeLab3D.m
