### 2019-08-15 #1173 @kyleniemeyer
@openjournals/joss-eics ok, this is ready for acceptance from my point of view
