### 2019-08-15 #1173 @pankajkmishra
Thanks @kyleniemeyer , and everyone :)
